Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -6,7 +6,7 @@ int main(){
     float rad;
     scanf("%f",&rad);
     int radi=2;
-    float power=pow(rad,radi);
+    int power=pow(rad,radi);
     float Area=3.14*power;
 
     printf("Area: %.2f",Area);
